Move Script into include/, src/ files
