+++ main.cpp
@@ -10,7 +10,7 @@
 #include "start_screen.hpp"
 #include "world_render.hpp"
 #include <world_navigation_gui/screen.hpp>
-#include "script.hpp"
+#include <script.hpp>
 
 #include "game/project.hpp"
 #include "game/load_game_scripts.hpp"
